SegmentedControlItemContainer created.
